Move pytest entrypoints to own module
